refactor: split tool-call-handler unit test tree
