+++ crates/assay-core/src/mcp/tool_call_handler/tests/mod.rs
@@ -1,0 +1,8 @@
+mod approval;
+mod classification;
+mod delegation;
+mod emission;
+mod fixtures;
+mod lifecycle;
+mod redaction;
+mod scope;
